refactor: Add UI Traits restriction for m2m integrations
