+++ ui/apps/platform/src/Containers/AccessControl/AccessScopes/AccessScopes.tsx
@@ -32,7 +32,7 @@ import AccessControlHeading from '../AccessControlHeading';
 import AccessControlBreadcrumbs from '../AccessControlBreadcrumbs';
 import AccessControlHeaderActionBar from '../AccessControlHeaderActionBar';
 import usePermissions from '../../../hooks/usePermissions';
-import { isUserResource } from '../traits';
+import { isUserResource } from '../../traits';
 
 const entityType = 'ACCESS_SCOPE';
