+++ ui/apps/platform/src/Containers/AccessControl/AuthProviders/AuthProviderForm.tsx
@@ -53,7 +53,7 @@ import {
 } from './authProviders.utils';
 import { AccessControlQueryAction } from '../accessControlPaths';
 import { TraitsOriginLabel } from '../TraitsOriginLabel';
-import { isUserResource } from '../traits';
+import { isUserResource } from '../../traits';
 
 export type AuthProviderFormProps = {
     isActionable: boolean;
