+++ ui/apps/platform/src/Containers/AccessControl/Roles/RolesList.tsx
@@ -9,7 +9,7 @@ import { PermissionSet, Role } from 'services/RolesService';
 import { AccessControlEntityLink } from '../AccessControlLinks';
 import { AccessControlQueryFilter } from '../accessControlPaths';
 import usePermissions from '../../../hooks/usePermissions';
-import { getOriginLabel, isUserResource } from '../traits';
+import { getOriginLabel, isUserResource } from '../../traits';
 
 // Return whether an auth provider rule refers to a role name,
 // therefore need to disable the delete action for the role.
