+++ ui/apps/platform/src/Containers/AccessControl/TraitsOriginLabel.tsx
@@ -1,7 +1,7 @@
 import React, { ReactElement } from 'react';
 import { Label } from '@patternfly/react-core';
 import { Traits } from 'types/traits.proto';
-import { getOriginLabel, originLabelColours } from './traits';
+import { getOriginLabel, originLabelColours } from '../traits';
 
 export type TraitsOriginLabelProps = {
     traits?: Traits;
